sandbox - move workspace tags to sandbox
