+++ src/vs/workbench/contrib/tags/electron-sandbox/workspaceTags.ts
